Move contexts into one folder
