+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/PylonBlockSchema.kt
@@ -1,5 +1,6 @@
 package io.github.pylonmc.pylon.core.block
 
+import io.github.pylonmc.pylon.core.block.context.BlockCreateContext
 import io.github.pylonmc.pylon.core.registry.PylonRegistry
 import io.github.pylonmc.pylon.core.util.findConstructorMatching
 import org.bukkit.Keyed
