+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/SimplePylonBlock.kt
@@ -1,5 +1,6 @@
 package io.github.pylonmc.pylon.core.block
 
+import io.github.pylonmc.pylon.core.block.context.BlockCreateContext
 import org.bukkit.block.Block
 import org.bukkit.persistence.PersistentDataContainer
