+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/base/BreakHandler.kt
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.core.block.base
 
-import io.github.pylonmc.pylon.core.block.BlockBreakContext
+import io.github.pylonmc.pylon.core.block.context.BlockBreakContext
 import org.bukkit.inventory.ItemStack
 
 interface BreakHandler {
