+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/block/context/BlockItemContext.kt
@@ -1,4 +1,4 @@
-package io.github.pylonmc.pylon.core.block
+package io.github.pylonmc.pylon.core.block.context
 
 import org.bukkit.entity.Player
