+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/event/PylonBlockBreakEvent.kt
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.core.event
 
-import io.github.pylonmc.pylon.core.block.BlockBreakContext
+import io.github.pylonmc.pylon.core.block.context.BlockBreakContext
 import io.github.pylonmc.pylon.core.block.PylonBlock
 import org.bukkit.block.Block
 import org.bukkit.event.Cancellable
