+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/item/base/BlockPlacer.kt
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.core.item.base
 
-import io.github.pylonmc.pylon.core.block.BlockCreateContext
+import io.github.pylonmc.pylon.core.block.context.BlockCreateContext
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema
 import io.github.pylonmc.pylon.core.persistence.blockstorage.BlockStorage
 import org.bukkit.block.Block
