+++ pylon-core/src/main/kotlin/io/github/pylonmc/pylon/core/persistence/blockstorage/PhantomBlock.kt
@@ -1,7 +1,7 @@
 package io.github.pylonmc.pylon.core.persistence.blockstorage
 
-import io.github.pylonmc.pylon.core.block.BlockCreateContext
-import io.github.pylonmc.pylon.core.block.BlockItemContext
+import io.github.pylonmc.pylon.core.block.context.BlockCreateContext
+import io.github.pylonmc.pylon.core.block.context.BlockItemContext
 import io.github.pylonmc.pylon.core.block.PylonBlock
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema
 import io.github.pylonmc.pylon.core.item.ItemStackBuilder
