+++ pylon-test/src/main/java/io/github/pylonmc/pylon/test/test/block/SimpleBlockWithSchemaTest.java
@@ -1,6 +1,6 @@
 package io.github.pylonmc.pylon.test.test.block;
 
-import io.github.pylonmc.pylon.core.block.BlockCreateContext;
+import io.github.pylonmc.pylon.core.block.context.BlockCreateContext;
 import io.github.pylonmc.pylon.core.block.PylonBlock;
 import io.github.pylonmc.pylon.core.block.PylonBlockSchema;
 import io.github.pylonmc.pylon.core.registry.PylonRegistry;
